requirements.txt
Flask==2.3.3
Flask-SQLAlchemy==3.0.5
celery==5.3.4
redis==5.0.1
twilio==8.10.0
python-dotenv==1.0.0
schedule==1.2.0
